+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Show source statistics in table when getting sources
+
 ## v0.18.2
 -Add ability to filter on user properties when getting comments